# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v2/auth/callback": {
      "get": {
        "description": "Handles Google OAuth callback, creates/updates user, generates JWT, and redirects to frontend with tokens.",
        "operationId": "AuthController_callback",
        "parameters": [],
        "responses": {
          "302": {
            "description": "Redirects to frontend with access_token in query params"
          },
          "401": {
            "description": "OAuth authentication failed"
          }
        },
        "summary": "Google OAuth callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/v2/auth/refresh": {
      "get": {
        "description": "Generate a new JWT token using a valid existing token. Token expires in 7 days.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New JWT token generated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "jwt_token_string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Refresh JWT token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v2/collections": {
      "get": {
        "description": "Retrieve collections based on visibility and ownership. Returns user's own collections plus public/unlisted ones. Supports cursor-based pagination.",
        "operationId": "CollectionsController_find_all",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order (default: desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort by field (default: created_at)",
            "schema": {
              "enum": [
                "created_at",
                "name"
              ],
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of collections"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all collections",
        "tags": [
          "collections"
        ]
      },
      "post": {
        "description": "Create a new collection with private, unlisted, or public visibility",
        "operationId": "CollectionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Collection data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "My Collection",
                  "user_id": "uuid",
                  "description": "Description",
                  "visibility": "private"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a collection",
        "tags": [
          "collections"
        ]
      }
    },
    "/v2/collections/{id}": {
      "get": {
        "description": "Retrieve a specific collection. Private collections require ownership.",
        "operationId": "CollectionsController_find_by_id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - No access to private collection"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get collection by ID",
        "tags": [
          "collections"
        ]
      },
      "put": {
        "description": "Update an existing collection. Requires ownership.",
        "operationId": "CollectionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated collection data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Updated Name",
                  "visibility": "public"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the owner"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a collection",
        "tags": [
          "collections"
        ]
      },
      "delete": {
        "description": "Remove a collection. Requires ownership.",
        "operationId": "CollectionsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Not the owner"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a collection",
        "tags": [
          "collections"
        ]
      }
    },
...
    "/v2/comments": {
      "get": {
        "description": "Retrieve comments with filtering, sorting, and pagination (offset or cursor-based). Supports threaded comments and text annotations.",
        "operationId": "CommentsController_find_all",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort order (default: desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "Sort by field (default: created_at)",
            "schema": {
              "enum": [
                "created_at",
                "content"
              ],
              "type": "string"
            }
          },
          {
            "name": "parent_comment_id",
            "required": false,
            "in": "query",
            "description": "Filter by parent comment UUID (get replies)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Filter by user UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": false,
            "in": "query",
            "description": "Filter by document UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Base64 encoded cursor for cursor-based pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (1-100, default: 25)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all comments",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "description": "Add a new comment to a document or reply to another comment. Supports text selection offsets for annotations.",
        "operationId": "CommentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Comment data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "user_id": "uuid",
                  "document_id": "uuid",
                  "content": "Great point!",
                  "start_offset": 0,
                  "end_offset": 100
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully"
          },
          "400": {
            "description": "Bad request - invalid offsets"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a comment",
        "tags": [
          "comments"
        ]
      }
    },
    "/v2/comments/{id}": {
      "put": {
        "description": "Update an existing comment",
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated comment data",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "content": "Updated comment text"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Comment not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a comment",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "description": "Remove a comment",
        "operationId": "CommentsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Comment not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a comment",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ohara API",
    "description": "REST API for Ohara App developed por EnSer management system with AI agent integration",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Admin API key for server-to-server authentication"
      }
    },
...
  }
}